requirements.txt
yt-dlp
groq
python-dotenv
ddgs
crewai
langchain-google-genai
streamlit
litellm